+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Release Candidate
 
+### Features
+
+- Implement a new option `dateTimeFormat` to change the time format for DateTime component (INDIGO Sprint 221125, [!376](https://github.com/TeskaLabs/asab-webui/pull/376))
+
 ## v22.48
 
 ### Breaking changes