+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- Implement a new option `dateTimeFormat` to change the time format for DateTime component (INDIGO Sprint 221125, [!376](https://github.com/TeskaLabs/asab-webui/pull/376))
+- Implement a new option `dateTimeFormat` to change the time format for DateTime component (INDIGO Sprint 221209, [!376](https://github.com/TeskaLabs/asab-webui/pull/376))
 
 ## v22.48