+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Release Candidate
 
+### Features
+
+- Implement a new option `dateTimeFormat` to change the time format for DateTime component (INDIGO Sprint 221209, [!376](https://github.com/TeskaLabs/asab-webui/pull/376))
+
 ### Refactoring
 
 - In TreeMenu component, display all closed folder nodes (even empty) with plus mark (`+`) to visually separate it from file nodes (INDIGO Sprint 221209, [!380](https://github.com/TeskaLabs/asab-webui/pull/380))